Update web_server
Please note that this change is so that the LED will turn on when the site is armed, not when an actual alarm is triggered. This was changed for testing purposes. 

To achieve the relay flipping based off an alarm being raised, just change the event type that we are parsing for.